+++ eno-core/src/main/java/fr/insee/eno/core/processing/out/steps/lunatic/LunaticSuggesterOptionResponses.java
@@ -3,7 +3,7 @@ package fr.insee.eno.core.processing.out.steps.lunatic;
 import fr.insee.eno.core.exceptions.business.InvalidSuggesterExpression;
 import fr.insee.eno.core.exceptions.technical.MappingException;
 import fr.insee.eno.core.processing.ProcessingStep;
-import fr.insee.eno.core.utils.VtlSyntaxUtils;
+import fr.insee.eno.core.utils.vtl.VtlSyntaxUtils;
 import fr.insee.lunatic.model.flat.*;
 import fr.insee.lunatic.model.flat.variable.CalculatedVariableType;
 import fr.insee.lunatic.model.flat.variable.CollectedVariableType;