+++ eno-core/src/main/java/fr/insee/eno/core/reference/EnoVariableIndex.java
@@ -1,0 +1,29 @@
+package fr.insee.eno.core.reference;
+
+import fr.insee.eno.core.model.EnoQuestionnaire;
+import fr.insee.eno.core.model.variable.Variable;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class EnoVariableIndex implements VariableIndex{
+    private final Map<String, Variable> index = new HashMap<>();
+
+    public void indexVariables(EnoQuestionnaire enoQuestionnaire){
+        enoQuestionnaire.getVariables().forEach(
+                variable -> index.put(variable.getId(), variable)
+        );
+    }
+
+    @Override
+    public Object get(String id) {
+        return index.get(id);
+    }
+
+    @Override
+    public boolean containsVariable(String variableName) {
+        return index.values().stream()
+                .map(Variable::getName)
+                .anyMatch(variableName::equals);
+    }
+}